Fix: Conversion Center - refactoring
